requirements.txt
mss>=9.0.0
Pillow>=10.0.0
PyQt6>=6.6.0
pynput>=1.7.6
httpx>=0.27.0
easyocr>=1.7.0
python-dotenv>=1.0.0
